requirements.txt
python-telegram-bot[webhooks]==21.2
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
